requirements.txt
Flask==2.2.5
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.2
elasticsearch==8.15.1
sentence-transformers==2.2.2
torch==2.5.0